requirements.txt
requests
numpy
python-dotenv
sentence-transformers
